Upgrade fresh to 0.5.2
This fixes the following vulnerability found in package-lock.json

CVE-2017-16119 More information
high severity
Vulnerable versions: < 0.5.2
Patched version: 0.5.2
Fresh is a module used by the Express.js framework for HTTP response
freshness testing. It is vulnerable to a regular expression denial of
service when it is passed specially crafted input to parse. This causes
the event loop to be blocked causing a denial of service condition.