Simplify config of base url the app points to
This decouples the code from the release channels. We only need to
specify the base url when making a build. I had to provide a default
because when using the Expo Go client in development there's no build. I
prefered this over requiring to pass an ENV var to `nxp expo start` or
setting up dir env. This is the scenario
https://docs.expo.dev/build-reference/variables/#can-i-share-environment-variables-defined-in
describes. Those solutions seemed too convoluted for our simple
duplication.

This makes the eas config look like a vanilla one, as stated in
https://docs.expo.dev/build/eas-json/#common-use-cases. This separates
the concept of "build profile" from "release channel", which let's us
combine them as we wish, without having to change the app's code. More
details about channels:
https://docs.expo.dev/eas-update/how-eas-update-works/#distributing-builds.