Consistently name build profiles and channels
According to
https://docs.expo.dev/eas-update/how-eas-update-works/#matching-updates-and-builds,
this will make sure "preview" builds end up in the app store's staging.
I have no clue how to check this though.

It avoids the need to get familiar with another concept; the build
profile. It's also highly unlikely that we have to combine different
profiles and channels.