Pass env vars to dynamic app configuration
It's not immediately obvious on
https://docs.expo.dev/guides/environment-variables/, but it turns out
that for ENV vars specified in the eas.json's profiles to be read from
`Constants.expoConfig.extra`, the app configuration needs to be dynamic
and set them from `process.env`.

This fixes the issue we were trying to solve on
https://github.com/coopdevs/timeoverflow-mobile-app/pull/14/files#diff-9b5b5955d0e684e7904f3fc143e07aaa11b7e7c3b36d10985c4aff2554940cbfR24.

Besides, using dynamic configuration gives more options to better adapt
it to our needs. Check https://docs.expo.dev/workflow/configuration/ for
details.